Move the sorting classes test
